pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mtuci-antiplagiarism"
version = "0.1.0"
description = "Локальная система анализа текстовых совпадений (антиплагиат)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "PySide6>=6.6.0",
    "python-docx>=1.1.0",
    "pypdf>=4.0.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "rapidfuzz>=3.0.0",
    "reportlab>=4.0.0",
]

[project.scripts]
antiplagiarism-gui = "gui_entry:main_gui"
antiplagiarism-cli = "gui_entry:main_cli"

[tool.setuptools]
py-modules = ["gui_entry"]

# PyInstaller: uv sync --group build && uv run pyinstaller Antiplagiarism.spec
[dependency-groups]
build = ["pyinstaller>=6.0"]
